OneDrive/Desktop/LLM-Project/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pypdf==5.0.0
python-docx==1.1.2
scikit-learn==1.5.2
numpy==2.1.1
httpx==0.27.2
python-dotenv==1.0.1
datasets==2.21.0
ollama==0.3.3
sentence-transformers==3.1.1
transformers>=4.44.0
groq==0.11.0
google-generativeai==0.8.3
pdfplumber==0.11.4
redis==5.2.1
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-prometheus==0.52b1
prometheus-client==0.21.1
